// src/models/CamContact.ts
export interface CamContact {
  id: number;
  oldContactId?: number;
  oldCompanyId?: number;
  gmaccountNum?: string;
  company?: string;
  contact?: string;
  companyId?: string;
  lastname?: string;
  title?: string;
  source?: string;
  dear?: string;
  assistant?: string;
  phoneMain?: string;
  extension?: string;
  phoneDirect?: string;
  phoneCell?: string;
  fax?: string;
  email?: string;
  website?: string;
  other1?: string;
  other2?: string;
  other3?: string;
  other4?: string;
  label1?: string;
  label2?: string;
  label3?: string;
  label4?: string;
  birthday?: string;
  birthDate?: Date;
  spouse?: string;
  kids?: string;
  interests?: string;
  address1?: string;
  address2?: string;
  address3?: string;
  city?: string;
  state?: string;
  zip?: string;
  country?: string;
  accountExec?: string;
  accountMgr?: string;
  accountCoord?: string;
  accountOther?: string;
  contactType?: string;
  salesTeam?: string;
  acctReps?: string;
  rwaccountNum?: string;
  poaccountNum?: string;
  activeStatus?: number;
  terms?: number;
  warrantyTerms?: number;
  eFlyer?: number;
  holidayGift?: string;
  gift1?: boolean;
  gift2?: boolean;
  competitor?: boolean;
  fnecard?: boolean;
  anccard?: boolean;
  mainVendor?: boolean;
  numOfSites?: number;
  entryDate?: Date;
  enteredBy?: string;
  modifiedDate?: Date;
  modifiedBy?: string;
  sContactIds?: string;
  sCompanyIds?: string;
  mfgs?: string;
  verizonArea?: string;
  verizonRegion?: string;
  verizonMarket?: string;
  btscount?: number;
  btstimeStamp?: Date;
  clauseId?: number;
  tempKatietoTimG?: number;
  generalInfo?: string;
}